Moved domain classes into their own files, added tests for these events. Added comments to remind adding tests
