+++ src/SelfService/Domain/Events/SchemaRegistered.cs
@@ -1,0 +1,8 @@
+using SelfService.Domain.Models;
+
+namespace SelfService.Domain.Events;
+
+public class SchemaRegistered : IDomainEvent
+{
+    public string? MessageContractId { get; set; }
+}
